requirements.txt
Faker
jsonpickle
mysql-connector-python-rf
pony
py
PyMySQL
pytest
pytest-bdd
python-dateutil
selenium
allure-pytest
allure-python-commons
